TMP double형 GC안나오게 추가대응
TMP double형 GC안나오게 추가대응

+++ PTT/Assets/Main/Scripts/FrameWork/Tables/ComboDataAsset.cs
@@ -6,7 +6,7 @@ namespace GameBerry
     [System.Serializable]
     public class ComboData
     {
-        public int ComboCount;
+        public long ComboCount;
         public int ConditionIndex;
     }
 
@@ -16,7 +16,7 @@ namespace GameBerry
         public List<ComboData> comboDatas = new List<ComboData>();
 
         //------------------------------------------------------------------------------------
-        public ComboData GetData(int count)
+        public ComboData GetData(long count)
             => comboDatas.Find(x => x.ComboCount == count);
         //------------------------------------------------------------------------------------
     }
